requirements.txt
flask
google-generativeai
gunicorn
python-dotenv